cloud-functions-fixed/qdrant-loader/requirements.txt
google-cloud-storage==2.14.0
google-cloud-pubsub==2.19.0
qdrant-client==1.7.0
functions-framework==3.5.0
